requirements.txt

python-telegram-bot==13.15
pillow
pymongo
requests
